+++ packages/moonstone/Scroller/ScrollableNative.js
@@ -7,6 +7,7 @@
 import clamp from 'ramda/src/clamp';
 import classNames from 'classnames';
 import {contextTypes as contextTypesResize} from '@enact/ui/Resizable';
+import {contextTypes as contextTypesRemeasurable} from '@enact/ui/Remeasurable';
 import {contextTypes as contextTypesRtl} from '@enact/i18n/I18nDecorator';
 import deprecate from '@enact/core/internal/deprecate';
 import {forward} from '@enact/core/handle';
@@ -211,7 +212,7 @@ const ScrollableHoC = hoc({configureSpotlight: false}, (config, Wrapped) => {
 			verticalScrollbar: 'auto'
 		}
 
-		static childContextTypes = contextTypesResize
+		static childContextTypes = Object.assign({}, contextTypesResize, contextTypesRemeasurable)
 		static contextTypes = contextTypesRtl
 
 		constructor (props) {
@@ -248,7 +249,8 @@ const ScrollableHoC = hoc({configureSpotlight: false}, (config, Wrapped) => {
 
 		getChildContext () {
 			return {
-				invalidateBounds: this.enqueueForceUpdate
+				invalidateBounds: this.enqueueForceUpdate,
+				remeasure: this.state.isVerticalScrollbarVisible
 			};
 		}
 
@@ -269,8 +271,13 @@ const ScrollableHoC = hoc({configureSpotlight: false}, (config, Wrapped) => {
 			configureSpotlightContainer(nextProps);
 		}
 
-		componentWillUpdate () {
+		componentWillUpdate (nextProps, nextState, nextContext) {
 			this.deferScrollTo = true;
+			if (nextContext.remeasure !== this.context.remeasure || nextState.isVerticalScrollbarVisible !== this.state.isVerticalScrollbarVisible) {
+				this.setState((prevState) => ({
+					remeasure: !prevState.remeasure
+				}));
+			}
 		}
 
 		componentDidUpdate () {